fix model issue

+++ src/train_full_model.py
@@ -90,7 +90,7 @@ optimizer = torch.optim.Adam(model.parameters(), lr=learning_rate)
 for epoch in range(num_epochs):
     for i in range(0, labels.shape[0], batch_size):
         batched_inputs = inputs[i: i + batch_size]
-        batched_labels = labels[i: i + batch_size].float()
+        batched_labels = labels[i: i + batch_size]
         # Move tensors to the configured device
         # inputs = inputs.reshape(-1, 28 * 28).to(device)
         # batched_labels = batched_labels.to(device)
@@ -144,7 +144,7 @@ with torch.no_grad():
 
     for i in range(0, labels.shape[0], batch_size):
         batched_inputs = inputs[i: i + batch_size]
-        batched_labels = labels[i: i + batch_size].float()
+        batched_labels = labels[i: i + batch_size]
         # images = images.reshape(-1, 28 * 28).to(device)
         outputs = model(batched_inputs)
         # _, predicted = torch.max(outputs.data, 1)
